requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Evaluation
nltk>=3.8.0
pycocoevalcap>=1.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.13.0
psutil>=5.9.0
pynvml>=11.5.0

# Development
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0
